1. Support unit test for Page. 2. Fix bug    1) Always seek to right position before read/write Page::m_body.    2) Always call OpenTempFile before read/write, and call CloseTempFile after it, when using tmp file.    3) Should call seekg(0) before call rdbuf() as the last rdbuf() could seek the get pos to the stream end.

+++ src/data/Cache.cpp
@@ -200,7 +200,7 @@ size_t Cache::Read(const string &fileId, off_t offset, size_t len, char *buffer,
     return page->Read(offset, sz, buffer);
   } else {  // Have Multipule pages.
     // read first page
-    auto readSize = page->Read(offset, buffer);
+    auto readSize = page->Read(static_cast<off_t>(offset), buffer);
     page = pagelist.front();
     pagelist.pop_front();
     // read middle pages
@@ -211,7 +211,7 @@ size_t Cache::Read(const string &fileId, off_t offset, size_t len, char *buffer,
     }
     // read last page
     auto sz = std::min(outcome.first - readSize, len - readSize);
-    readSize += page->Read(sz, buffer + readSize);
+    readSize += page->Read(static_cast<size_t>(sz), buffer + readSize);
     return readSize;
   }
 }
